no solution. some minor debugging and fix attempts

+++ Spaceport/23-24/Code/Teensy-Based Avionics/src/AprilFilter.h
@@ -3,11 +3,10 @@
 
 #include "Kalman_Filter.h"
 
-Matrix get_F(int dt);
-Matrix get_G(int dt);
+Matrix get_F(double dt);
+Matrix get_G(double dt);
 Matrix get_H(int has_gps, int has_barometer);
 LinearKalmanFilter *initializeFilter();
-double* iterateFilter(LinearKalmanFilter kf, int dt, double* input, double* measurement, int has_gps, int has_barometer);
-
+double *iterateFilter(LinearKalmanFilter kf, double dt, double *input, double *measurement, int has_gps, int has_barometer);
 
 #endif
